search functionality streamlined
