+++ src/components/home/Home.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { Label } from "reactstrap";
 import NavBar from "../nav/NavBar";
 import MedicationFormModal from "../medication/MedicationFormModal";
@@ -8,23 +8,6 @@ const Home = (props) => {
     const { drugs, renderWidget, imageDesc, removeDrug, isChecked, drugImage, handleChange, getIdOfDrug, drug, handlePharmacyDropdown, pharmacyList, uploadImage, handleFieldChange, handleDrugForm, toggle, modal, toggleNested, toggleAll, nestedModal, closeAll } = props;
     const hasUser = props.hasUser
     const clearUser = props.clearUser
-
-    const [filteredDrugsArray, setFilteredDrugsArray] = useState([])
-
-    const getMatchingCards = (event) => {
-        let searchEvent = event.target.value
-        let filteringDrugsArray = drugs.filter(drug => {
-            let drugValues = Object.values(drug)
-            for (let i = 0; i < drugValues.length; i++) {
-                return drugValues.join().toLowerCase().includes(searchEvent.toLowerCase())
-            }
-        })
-        if (searchEvent === "") {
-            filteringDrugsArray = []
-        }
-        setFilteredDrugsArray(filteringDrugsArray)
-    }
-
 
     return (
         <>
@@ -37,7 +20,7 @@ const Home = (props) => {
                     <Label htmlFor="addMedication-image"><h5>Add New Medication</h5></Label>
                 </div>
 
-                <SearchBar getMatchingCards={getMatchingCards} filteredDrugsArray={filteredDrugsArray} removeDrug={removeDrug} getIdOfDrug={getIdOfDrug} handleChange={handleChange} isChecked={isChecked} />
+                <SearchBar drugs={drugs} removeDrug={removeDrug} getIdOfDrug={getIdOfDrug} handleChange={handleChange} isChecked={isChecked} />
                 <MedicationFormModal imageDesc={imageDesc} renderWidget={renderWidget} drug={drug} handlePharmacyDropdown={handlePharmacyDropdown} pharmacyList={pharmacyList} drugImage={drugImage} uploadImage={uploadImage} handleFieldChange={handleFieldChange} handleDrugForm={handleDrugForm}
                     getIdOfDrug={getIdOfDrug} nestedModal={nestedModal} toggle={toggle} modal={modal} toggleNested={toggleNested} toggleAll={toggleAll} closeAll={closeAll} isChecked={isChecked} />
             </div>
