+++ src/components/search/SearchBar.js
@@ -1,12 +1,26 @@
-import React, { useState, useEffect } from "react";
-import ApplicationManager from "../modules/ApplicationManager";
+import React, { useState } from "react";
 import MedicationCard from "../medication/MedicationCard";
 import { Input, Row, Col } from "reactstrap"
 import "./styles/SearchBar.css"
 
 
 const SearchBar = (props) => {
-    const { filteredDrugsArray, getMatchingCards } = props
+    const { drugs } = props
+    const [filteredDrugsArray, setFilteredDrugsArray] = useState([])
+
+    const getMatchingCards = (e) => {
+        let searchTerm = e.target.value
+        let filteringDrugsArray = drugs.filter(drug => {
+            let drugValues = Object.values(drug)
+            for (let i = 0; i < drugValues.length; i++) {
+                return drugValues.join().toLowerCase().includes(searchTerm.toLowerCase())
+            }
+        })
+        if (searchTerm === "") {
+            filteringDrugsArray = [];
+        }
+        setFilteredDrugsArray(filteringDrugsArray)
+    }
 
     return (
         <>
@@ -17,7 +31,7 @@ const SearchBar = (props) => {
             </div>
 
             <div className="searchBar-result-overlay">
-                {(window.location.pathname === "/" && filteredDrugsArray) && filteredDrugsArray.map(drug => {
+                {filteredDrugsArray && filteredDrugsArray.map(drug => {
                     return (
 
                         <Row className="div-medicationCard-searchResult">
@@ -25,14 +39,13 @@ const SearchBar = (props) => {
                                 <MedicationCard drugId={drug.id} {...props} drug={drug} />
                             </Col>
                         </Row>
-
                     )
-
                 })}
             </div>
 
         </>
     )
+
 }
 
 export default SearchBar
